Addressed CR comments from previous revision.
